Time: 3 ms (95.2%), Space: 10.2 MB (89.05%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     int change(int amount, vector<int>& coins) {
-        vector<uint64_t> dp(amount + 1, 0);
+        vector<unsigned long long> dp(amount + 1, 0);
         dp[0] = 1;
         for (int coin : coins) {
             for (int s = coin; s < dp.size(); s++) {
                 dp[s] += dp[s - coin];
             }
         }
-        return dp.back();
+        return (int)dp.back();
     }
 };
 /*
